refactor: HQL Function 분리
